+++ src/Raven.CodeAnalysis/Compilation.cs
@@ -3800,7 +3800,7 @@ public partial class Compilation
             SpecialType.System_Runtime_CompilerServices_AsyncStateMachineAttribute => "System.Runtime.CompilerServices.AsyncStateMachineAttribute",
             SpecialType.System_Runtime_CompilerServices_IteratorStateMachineAttribute => "System.Runtime.CompilerServices.IteratorStateMachineAttribute",
             SpecialType.System_Threading_Tasks_Task => "System.Threading.Tasks.Task",
-            SpecialType.System_Threading_Tasks_Task_T => "System.Threading.Tasks.Task`1",
+            SpecialType.System_Threading_Tasks_Task_T => Options.UseHeapAsyncStateMachines && Options.TargetCoreAssemblyName is not null ? "System.Tasks.Task`1" : "System.Threading.Tasks.Task`1",
             SpecialType.System_Runtime_InteropServices_WindowsRuntime_EventRegistrationToken => "System.Runtime.InteropServices.WindowsRuntime.EventRegistrationToken",
             SpecialType.System_Runtime_InteropServices_WindowsRuntime_EventRegistrationTokenTable_T => "System.Runtime.InteropServices.WindowsRuntime.EventRegistrationTokenTable`1",
             SpecialType.System_ValueTuple_T1 => "System.ValueTuple`1",